% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cat_observations}
\alias{cat_observations}
\title{GTA Cat Observations}
\format{
\subsection{\code{cat_observations}}{

A data frame with 204 rows and 6 variables:
\describe{
\item{incident_id}{Unique identifier for the incident}
\item{note_id}{Unique identifier for the submitted incident note}
\item{date_received}{Date the note was submitted}
\item{cat_name}{Primary cat  involved in the incident}
\item{neighborhood}{Neighborhood where the incident was observed}
\item{observation_note}{Description of the incident}
}
}
}
\source{
Generated with the help of Co-pilot; edited and finalized by a human
}
\usage{
cat_observations
}
\description{
Fictional data set of cat observations in the Greater Toronto Area.
}
\keyword{datasets}
