% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_fiscal_year.R
\name{gc_fiscal_year}
\alias{gc_fiscal_year}
\title{Convert dates to Government of Canada fiscal year}
\usage{
gc_fiscal_year(date, fy_format = c("long", "short"))
}
\arguments{
\item{date}{A date vector coercible to Date.}

\item{fy_format}{long (YYYY-YYYY) or short (YY-YY) format?}
}
\value{
A character vector of fiscal years in the format
"YYYY-YYYY" or "YY-YY"
}
\description{
Government of Canada fiscal years begin on April 1 and end on
March 31 of the following year.

This function is meant to show best practices for writing an R function.
See \code{\link[=fy]{fy()}} for a counter example.
}
\examples{
gc_fiscal_year("2026-03-31", fy_format = "long")
gc_fiscal_year(c("2026-01-01", "2026-04-20"), fy_format = "short")

}
