% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_search.R
\name{keyword_search}
\alias{keyword_search}
\title{Keycode search and filter}
\usage{
keyword_search(KEYWORDS_DF, DATA, SEARCH_COLUMN)
}
\arguments{
\item{KEYWORDS_DF}{dataframe with columns KEYWORD and KEYCODE, generated from \code{\link[=mk_keyword_search_regex]{mk_keyword_search_regex()}}}

\item{DATA}{dataframe that includes a column with content to search}

\item{SEARCH_COLUMN}{string; name of DATA column in which to conduct the search}
}
\value{
returns DATA dataframe filtered for only those rows with keycode hits
}
\description{
Conduct a keycode search of a particular column of an input dataframe and
return filters that dataframe filtered for keycode hits.
Expects KEYCODE to be regex formatted strings
}
\examples{
data(cat_observations)
Search_Terms <- data.frame(
    KEYWORD = c("Mr Jones", "meow", "yowl"),
    KEYCODE = c("(?i)\\\\b(Mister|Mr\\\\.?) Jones?\\\\b", "(?i)meow", "(?i)yowl")
)
search_results <- keyword_search(
  KEYWORDS_DF = Search_Terms,
  DATA = cat_observations,
  SEARCH_COLUMN = "observation_note")

}
