% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_search_params.R
\name{wb_search_params}
\alias{wb_search_params}
\title{Create a search parameter tab in an openxlsx workbook}
\usage{
wb_search_params(wb, wb_style, params, sheet_num = 1)
}
\arguments{
\item{wb}{openxlsx wb object}

\item{wb_style}{openxlsx wb style list generated from \code{\link[=wb_style]{wb_style()}}}

\item{params}{list of search parameters including: title, start_date,
end_date, search_terms, and search_fields (see example)}

\item{sheet_num}{the sheet number to output the search parameters}
}
\value{
openxlsx wb object with a search parameter sheet added
}
\description{
Function to create a standardized search parameter tab that outlines
time range, fields, and search terms used in report. Use with \code{\link[=wb_style]{wb_style()}}
and \code{\link[=wb_summary_tab]{wb_summary_tab()}}
}
\examples{
 \dontrun{
search_params <- list(
  title = "Cat Search - Mr Jones",
  start_date = "2026-01-01",
  end_date = Sys.Date(),
  search_terms = data.frame(Search_Terms = c("Mr Jones", "meow", "yowl")),
  search_fields = data.frame(Search_Fields = "observation_note")
)

library(openxlsx)
wb <- createWorkbook()
addWorksheet(wb, sheetName = "Search Parameters", gridLines = FALSE)
wb_search_params(wb = wb, wb_style = wb_style(), params = search_params, sheet_num = 1)
saveWorkbook(wb, paste0(search_params$title, "_" , Sys.Date(), ".xlsx"), overwrite = TRUE)
}
}
