% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_style.R
\name{wb_style}
\alias{wb_style}
\title{openxlsx workbook style helper}
\usage{
wb_style()
}
\value{
list of openxlsx styling
}
\description{
Function to apply standard formatting to datatables created with openxlsx.
Use with \code{\link[=wb_search_params]{wb_search_params()}} and \code{\link[=wb_summary_tab]{wb_summary_tab()}}
}
\examples{
wb_style()

}
