% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_keyword_search_regex.R
\name{mk_keyword_search_regex}
\alias{mk_keyword_search_regex}
\title{Turn search terms into regex}
\usage{
mk_keyword_search_regex(
  KEYWORDS,
  STANDALONE = TRUE,
  IGNORE_CASE = TRUE,
  MAKE_PLURAL = FALSE
)
}
\arguments{
\item{KEYWORDS}{Character vector of search terms.}

\item{STANDALONE}{Logical. Should word boundaries be added? default = TRUE}

\item{IGNORE_CASE}{Logical. Should matching be case-insensitive? default = TRUE}

\item{MAKE_PLURAL}{Logical. Should an optional trailing "s" be added? default = FALSE}
}
\value{
A data frame with columns KEYWORD and KEYCODE (regex)
}
\description{
Takes a character vector and converts it to a dataframe containing
regex patterns suitable for use with keyword searches.
}
\details{
Creates a dataframe compatible for use with \code{\link[=keyword_search]{keyword_search()}}
}
\examples{
KEYWORDS_DF <- mk_keyword_search_regex(KEYWORDS = c("Pixel", "Beach(es)?"),
                                STANDALONE = TRUE, IGNORE_CASE = TRUE, MAKE_PLURAL = FALSE)

}
