% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_keyword_search_regex.R
\name{mk_keyword_search_regex}
\alias{mk_keyword_search_regex}
\title{Turn search terms into regex}
\usage{
mk_keyword_search_regex(
  KEYWORDS,
  STANDALONE = TRUE,
  IGNORE_CASE = TRUE,
  MAKE_PLURAL = FALSE
)
}
\arguments{
\item{KEYWORDS}{a string vector of search terms}

\item{STANDALONE}{boolean; standalone word/word boundaries? default = TRUE}

\item{IGNORE_CASE}{boolean; ignore upper/lower case? default = TRUE}

\item{MAKE_PLURAL}{boolean; make keyword plural by adding "s" to the end? default = FALSE}
}
\value{
dataframe of keywords and the corresponding regex keycodes
}
\description{
Takes a string vector and converts it to a dataframe with regex formatting
to handle word boundaries, case sensitivity, and basic pluralization.
Creates a dataframe compatible for use with \code{\link[=keyword_search]{keyword_search()}}
}
\examples{
KEYWORDS_DF <- mk_keyword_search_regex(c("Pixel", "Beach(es)?"),
                                STANDALONE = TRUE, IGNORE_CASE = TRUE, MAKE_PLURAL = FALSE)

}
